+++ src/_shared/ToggleNode.js
@@ -37,10 +37,6 @@ class ToggleNode {
     }
     this.node.addEventListener(ToggleNode.END_EVENT, onTransitionEnd)
   }
-
-  setNode(node) {
-    this.node = node
-  }
 }
 
 export default ToggleNode