+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Preview
 
 - SearchInput 新增可控属性 value
+- Modal 渲染到 body 节点内，样式及动画性能优化，修复嵌套 Modal 的样式问题
 - 因 [pace](http://github.hubspot.com/pace/docs/welcome/) 加载进度不够精确，文档站页面切换进度条升级到 [nprocess](http://ricostacruz.com/nprogress/)，与 react-router 配合使用